modele: Move Joueur into a new joueur package
