+++ src/main/java/fr/nc0/cda/modele/joueur/Joueur.java
@@ -4,7 +4,7 @@
  * that can be found in the COPYRIGHT file.
  */
 
-package fr.nc0.cda.modele;
+package fr.nc0.cda.modele.joueur;
 
 /** Représente une partie d'un jeu. */
 public class Joueur {
